7.11. Отрисуй меня полностью (часть 2) Исправил мелкие замечания. Добавил addEventListener. Перенёс библиотеку и удалил пустой файл.
