+++ js/gallery.js
@@ -1,21 +1,24 @@
 import {showBigPicture} from './fullsize-picture.js';
-import {isEscapeKey} from './util.js';
-import {hideBigPicture} from './fullsize-picture.js';
 
-const handlerPicture = (photos) => {
-  const picture = document.querySelectorAll('.pictures .picture');
-  picture.forEach((elem, index) => {
-    elem.addEventListener('click', () => showBigPicture(photos[index]));
+const template = document.querySelector('#picture').content;
+const templatePicture = template.querySelector('.picture');
+const pictureList = document.querySelector('.pictures');
+
+
+const renderPhotos = (photosData) => {
+  const templateFragment = document.createDocumentFragment();
+  photosData.forEach(({url, likes, comments, description}) => {
+    const result = {url, likes, comments, description};
+    const photoElement = templatePicture.cloneNode(true);
+    photoElement.querySelector('.picture__img').src = url;
+    photoElement.querySelector('.picture__likes').textContent = likes;
+    photoElement.querySelector('.picture__comments').textContent = comments.length;
+    templateFragment.append(photoElement);
+    photoElement.addEventListener('click', () => showBigPicture(result));
   });
-  const closeButton = document.querySelector('.big-picture__cancel');
-  closeButton.addEventListener('click', () => hideBigPicture());
-  document.addEventListener('keydown', (evt) => {
-    if (isEscapeKey(evt)) {
-      hideBigPicture();
-    }
-  });
+  pictureList.append(templateFragment);
 };
 
 export {
-  handlerPicture
+  renderPhotos
 };
