+++ js/main.js
@@ -1,10 +1,7 @@
-import {renderPhotos} from './miniature.js';
+import {renderPhotos} from './gallery.js';
 import {createPhotos, PHOTO_COUNT} from './data.js';
-import {handlerPicture} from './gallery.js';
-
 
 const photos = createPhotos(PHOTO_COUNT);
 
 renderPhotos(photos);
-handlerPicture(photos);
