+++ js/util.js
@@ -1,3 +1,5 @@
+import {hideBigPicture} from './fullsize-picture.js';
+
 /**
  * Функция, возвращающая случайное целое число из переданного диапазона включительно.
  *
@@ -6,7 +8,6 @@
  *
  */
 const getRandomInt = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
-
 /**
  * Проверка длинны строки.
  *
@@ -15,13 +16,19 @@ const getRandomInt = (min, max) => Math.floor(Math.random() * (max - min + 1)) +
  */
 const checkStringLength = (string, maxLength = 140) => string.length <= maxLength;
 
-
 const getRandomArrayElement = (elements) => elements[getRandomInt(0, elements.length - 1)];
 
 const isEscapeKey = (evt) => evt.key === 'Escape';
 
+const onPopupEscKeyDown = (evt) => {
+  if (isEscapeKey(evt)) {
+    evt.preventDefault();
+  }
+  hideBigPicture();
+};
+
 export {
   getRandomInt,
   getRandomArrayElement,
-  isEscapeKey
+  onPopupEscKeyDown
 };
